Cleaned unnecessary files and moved files to proper locations
