+++ .gitignore
@@ -4,3 +4,4 @@ code/venv
 code/__pycache__
 code/hurriscan.db
 venv
+hurriscan.db
