+++ code/routing/main.py
@@ -16,5 +16,5 @@ def hello_world():
 
 
 if __name__ == "__main__":
-    port = 7070
+    port = 5000
     app.run(host="0.0.0.0", port=port)
